+++ Pod/Classes/TBStateMachine.h
@@ -35,7 +35,7 @@
  *  - call -tearDown to deactivate the state machine
  *
  *  **Concurrency:**
- *  Event handlers, enter and exit handlers will be executed on a background queue.
+ *  Actions, guards, enter and exit blocks will be executed on a background queue.
  *  Make sure the code in these blocks is dispatched back onto the right queue.
  */
 @interface TBStateMachine : NSObject <TBStateMachineNode>
@@ -109,8 +109,21 @@
  */
 - (void)setInitialState:(id<TBStateMachineNode>)initialState;
 
+/**
+ *  Schedules an event.
+ *  The state machine will queue all events it receives until processing of the current state has finished.
+ *
+ *  @param event The given `TBStateMachineEvent` instance.
+ */
 - (void)scheduleEvent:(TBStateMachineEvent *)event;
 
+/**
+ *  Schedules an event with a given payload.
+ *  The state machine will queue all events it receives until processing of the current state has finished.
+ *
+ *  @param event The given `TBStateMachineEvent` instance.
+ *  @param data  The payload data.
+ */
 - (void)scheduleEvent:(TBStateMachineEvent *)event data:(NSDictionary *)data;
 
 @end